feat: 지도상에서 휠로 위치이동시에도 주소 변경되도록 추가, 검색기록 컴포넌트 분리 #16
