+++ src/components/SearchPage.tsx
@@ -3,28 +3,13 @@
 import styles from "@/styles/layout/_searchPage.module.scss"
 import { useState } from "react";
 import Topper from "@/components/SubTopper";
+import { SearchLog } from "./SearchLog";
 
 export default function Page() {
   const [inputCollectionSearch, setInputCollectionSearch] = useState("");
 
   const onChangeCollection = (e: any) => {
     setInputCollectionSearch(e.target.value);
-  }
-
-  function SearchLog({ searchString, searchCategory }: { searchString: string; searchCategory: string }) {
-    if(searchString.length >= 7)
-    {
-      searchString = searchString.substring(0, 7);
-      searchString = searchString + "...";
-    }
-    return(
-      <button>
-        <div className={styles.searchLog}>
-          <p>{searchString}</p>
-          <p className={styles.searchCategory}>{searchCategory}</p>
-        </div>
-      </button>
-    );
   }
 
   return(
